requirements.txt
ipywidgets
scikit-learn
matplotlib
seaborn
torch==2.4.1
openai==1.59.3
transformers==4.46.3
datasets==3.1.0
httpx==0.26.0
haversine
fastapi
pandas
uvicorn[standard]
langchain
langchain-openai
langchain-community
numpy==1.26.4
supabase==2.11.0
python-dotenv
chromadb